Fix narrowing conversions, dead store, and multilevel pointer casts
- bridge.c: Cast tolower() result to (char) to fix narrowing conversion
- bridge.c: Replace dead store with (void) cast for snprintf return value
- context.c: Cast tv_sec/tv_nsec and total_executions to double explicitly
- ruleset.c: Add explicit void*/T** casts for calloc, malloc, and free calls
  to silence bugprone-multi-level-implicit-pointer-conversion warnings

+++ src/vm/context.c
@@ -34,7 +34,7 @@ cdsl_get_time_us_internal(void)
 {
 	struct timespec ts;
 	clock_gettime(CLOCK_MONOTONIC, &ts);
-	return ts.tv_sec * 1e6 + ts.tv_nsec / 1e3;
+	return (double)ts.tv_sec * 1e6 + (double)ts.tv_nsec / 1e3;
 }
 
 cdsl_context_entry_t*
@@ -590,7 +590,7 @@ cdsl_vm_get_stats(const cdsl_vm_t* vm)
 	s->total_time_us = vm->stats.total_time_us;
 	s->avg_time_us = vm->stats.avg_time_us;
 	if (s->total_executions > 0) {
-		s->avg_time_us = s->total_time_us / s->total_executions;
+		s->avg_time_us = s->total_time_us / (double)s->total_executions;
 	}
 	return s;
 }
